+++ NEWS.md
@@ -5,6 +5,15 @@ editor_options:
     wrap: 72
 ---
 
+# photobiology 0.13.1
+
+* Remove functions `get_peaks()` and `get_valleys()` as their use has been 
+deprecated some time ago.
+* Rewrite function `find_peaks()` fixing a bug in the logic of threshold tests.
+This modifies the behaviour compared to version 0.13.1, when first introduced.
+* Add function `find_valleys()` and edit `valleys()` methods to use it.
+* These changes, repair the behaviour of `peaks()` and `valleys()` methods. 
+
 # photobiology 0.13.0
 
 This update focus is on 1) more efficient storage of metadata in attributes, 2)